feat : refactor DigitalEqualOperation place
