+++ kernels/jvm/src/main/java/io/github/mosser/arduinoml/kernel/samples/MultipleTransitionsTest.java
@@ -6,7 +6,7 @@ import io.github.mosser.arduinoml.kernel.generator.ToWiring;
 import io.github.mosser.arduinoml.kernel.generator.Visitor;
 
 import io.github.mosser.arduinoml.kernel.structural.actuators.DigitalActuator;
-import io.github.mosser.arduinoml.kernel.structural.expressions.digitalbinaryoperations.DigitalEqualOperation;
+import io.github.mosser.arduinoml.kernel.structural.expressions.DigitalEqualOperation;
 import io.github.mosser.arduinoml.kernel.structural.sensors.DigitalSensor;
 import io.github.mosser.arduinoml.kernel.structural.sensors.AnalogSensor;
 import io.github.mosser.arduinoml.kernel.structural.signals.*;
